+++ liblangutil/EVMVersion.h
@@ -64,6 +64,7 @@ public:
 	static EVMVersion shanghai() { return {Version::Shanghai}; }
 	static EVMVersion cancun() { return {Version::Cancun}; }
 	static EVMVersion prague() { return {Version::Prague}; }
+	static EVMVersion osaka() { return {Version::Osaka}; }
 
 	static std::vector<EVMVersion> allVersions() {
 		return {
@@ -80,6 +81,7 @@ public:
 			shanghai(),
 			cancun(),
 			prague(),
+			osaka(),
 		};
 	}
 
@@ -91,8 +93,10 @@ public:
 		return std::nullopt;
 	}
 
+	static EVMVersion firstWithEOF() { return {Version::Osaka}; }
+
 	bool isExperimental() const {
-		return m_version == Version::Prague;
+		return *this > EVMVersion{};
 	}
 
 	bool operator==(EVMVersion const& _other) const { return m_version == _other.m_version; }
@@ -115,6 +119,7 @@ public:
 		case Version::Shanghai: return "shanghai";
 		case Version::Cancun: return "cancun";
 		case Version::Prague: return "prague";
+		case Version::Osaka: return "osaka";
 		}
 		util::unreachable();
 	}
@@ -134,6 +139,7 @@ public:
 	bool hasBlobHash() const { return *this >= cancun(); }
 	bool hasMcopy() const { return *this >= cancun(); }
 	bool supportsTransientStorage() const { return *this >= cancun(); }
+	bool supportsEOF() const { return *this >= firstWithEOF(); }
 
 	bool hasOpcode(evmasm::Instruction _opcode, std::optional<uint8_t> _eofVersion) const;
 
@@ -155,7 +161,8 @@ private:
 		Paris,
 		Shanghai,
 		Cancun,
-		Prague
+		Prague,
+		Osaka,
 	};
 
 	EVMVersion(Version _version): m_version(_version) {}