+++ libyul/backends/evm/EVMDialect.cpp
@@ -155,14 +155,12 @@ std::set<std::string, std::less<>> createReservedIdentifiers(langutil::EVMVersio
 
 	auto eofIdentifiersException = [&](evmasm::Instruction _instr) -> bool
 	{
-		solAssert(!_eofVersion.has_value() || _evmVersion >= langutil::EVMVersion::prague());
-		return
-			!_eofVersion.has_value() &&
-			(
-				_instr == evmasm::Instruction::EXTCALL ||
-				_instr == evmasm::Instruction::EXTSTATICCALL ||
-				_instr == evmasm::Instruction::EXTDELEGATECALL
-			);
+		solAssert(!_eofVersion.has_value() || (*_eofVersion == 1 && _evmVersion.supportsEOF()));
+		if (_eofVersion.has_value())
+			// For EOF every instruction is reserved identifier.
+			return false;
+		return langutil::EVMVersion::firstWithEOF().hasOpcode(_instr, 1) &&
+			!langutil::EVMVersion::firstWithEOF().hasOpcode(_instr, std::nullopt);
 	};
 
 	std::set<std::string, std::less<>> reserved;
@@ -382,50 +380,51 @@ std::vector<std::optional<BuiltinFunctionForEVM>> createBuiltins(langutil::EVMVe
 	}
 	else // EOF context
 	{
-		builtins.emplace_back(createFunction(
-			"auxdataloadn",
-			1,
-			1,
-			EVMDialect::sideEffectsOfInstruction(evmasm::Instruction::DATALOADN),
-			ControlFlowSideEffects::fromInstruction(evmasm::Instruction::DATALOADN),
-			{LiteralKind::Number},
-			[](
-				FunctionCall const& _call,
-				AbstractAssembly& _assembly,
-				BuiltinContext&
-			) {
-				yulAssert(_call.arguments.size() == 1);
-				Literal const* literal = std::get_if<Literal>(&_call.arguments.front());
-				yulAssert(literal, "");
-				yulAssert(literal->value.value() <= std::numeric_limits<uint16_t>::max());
-				_assembly.appendAuxDataLoadN(static_cast<uint16_t>(literal->value.value()));
-			}
-		));
-
-		builtins.emplace_back(createFunction(
-			"eofcreate",
-			5,
-			1,
-			EVMDialect::sideEffectsOfInstruction(evmasm::Instruction::EOFCREATE),
-			ControlFlowSideEffects::fromInstruction(evmasm::Instruction::EOFCREATE),
-			{LiteralKind::String, std::nullopt, std::nullopt, std::nullopt, std::nullopt},
-			[](
-				FunctionCall const& _call,
-				AbstractAssembly& _assembly,
-				BuiltinContext& context
-			) {
-				yulAssert(_call.arguments.size() == 5);
-				Literal const* literal = std::get_if<Literal>(&_call.arguments.front());
-				auto const formattedLiteral = formatLiteral(*literal);
-				yulAssert(!util::contains(formattedLiteral, '.'));
-				auto const* containerID = valueOrNullptr(context.subIDs, formattedLiteral);
-				yulAssert(containerID != nullptr);
-				yulAssert(*containerID <= std::numeric_limits<AbstractAssembly::ContainerID>::max());
-				_assembly.appendEOFCreate(static_cast<AbstractAssembly::ContainerID>(*containerID));
-			}
+		if (_objectAccess)
+		{
+			builtins.emplace_back(createFunction(
+				"auxdataloadn",
+				1,
+				1,
+				EVMDialect::sideEffectsOfInstruction(evmasm::Instruction::DATALOADN),
+				ControlFlowSideEffects::fromInstruction(evmasm::Instruction::DATALOADN),
+				{LiteralKind::Number},
+				[](
+					FunctionCall const& _call,
+					AbstractAssembly& _assembly,
+					BuiltinContext&
+				) {
+					yulAssert(_call.arguments.size() == 1);
+					Literal const* literal = std::get_if<Literal>(&_call.arguments.front());
+					yulAssert(literal, "");
+					yulAssert(literal->value.value() <= std::numeric_limits<uint16_t>::max());
+					_assembly.appendAuxDataLoadN(static_cast<uint16_t>(literal->value.value()));
+				}
 			));
 
-		if (_objectAccess)
+			builtins.emplace_back(createFunction(
+				"eofcreate",
+				5,
+				1,
+				EVMDialect::sideEffectsOfInstruction(evmasm::Instruction::EOFCREATE),
+				ControlFlowSideEffects::fromInstruction(evmasm::Instruction::EOFCREATE),
+				{LiteralKind::String, std::nullopt, std::nullopt, std::nullopt, std::nullopt},
+				[](
+					FunctionCall const& _call,
+					AbstractAssembly& _assembly,
+					BuiltinContext& context
+				) {
+					yulAssert(_call.arguments.size() == 5);
+					Literal const* literal = std::get_if<Literal>(&_call.arguments.front());
+					auto const formattedLiteral = formatLiteral(*literal);
+					yulAssert(!util::contains(formattedLiteral, '.'));
+					auto const* containerID = valueOrNullptr(context.subIDs, formattedLiteral);
+					yulAssert(containerID != nullptr);
+					yulAssert(*containerID <= std::numeric_limits<AbstractAssembly::ContainerID>::max());
+					_assembly.appendEOFCreate(static_cast<AbstractAssembly::ContainerID>(*containerID));
+				}
+				));
+
 			builtins.emplace_back(createFunction(
 				"returncontract",
 				3,
@@ -449,6 +448,7 @@ std::vector<std::optional<BuiltinFunctionForEVM>> createBuiltins(langutil::EVMVe
 					_assembly.appendReturnContract(static_cast<AbstractAssembly::ContainerID>(*containerID));
 				}
 			));
+		}
 	}
 	yulAssert(
 		ranges::all_of(builtins, [](std::optional<BuiltinFunctionForEVM> const& _builtinFunction){